ai-agents-core/requirements.txt
# Web & API
fastapi
uvicorn
python-multipart

# Agent Orchestration
crewai
crewai[tools]
langchain
langchain-community
pydantic>=2.0.0

# Infrastructure & Tracing
redis
arize-phoenix
opentelemetry-api
opentelemetry-sdk

# Tools
python-dotenv
requests